refactor(api): uniform Long bank amounts + fuller GuildLookup test coverage
Address CodeRabbit on #62:
- GuildLookup bank amounts are now Long for a consistent API; GuildLookupImpl
  converts to LumaGuilds' Int-bounded BankService and rejects non-positive /
  out-of-range amounts (overflow guard moved into the impl). KDoc documents the
  Int ceiling. (Full Long-end-to-end BankService change is out of scope here.)
- Split bundled isMember/bank tests into focused cases; add coverage for
  hasShopPermission swallowing service exceptions, and hasRankAtLeast lower-
  authority + missing-target-rank false paths.

+++ src/main/kotlin/net/lumalyte/lg/api/GuildLookup.kt
@@ -50,11 +50,20 @@ interface GuildLookup {
     /** Guild bank balance for [guildId], or 0 if unknown. */
     fun getBankBalance(guildId: UUID): Long
 
-    /** Withdraw [amount] from [guildId]'s bank as [actorId]. True on success. */
-    fun bankWithdraw(guildId: UUID, actorId: UUID, amount: Int, reason: String): Boolean
+    /**
+     * Withdraw [amount] from [guildId]'s bank as [actorId]. True on success.
+     *
+     * Amounts are `Long` for a uniform API, but LumaGuilds' bank is Int-bounded
+     * internally, so amounts that are non-positive or exceed `Int.MAX_VALUE`
+     * (~2.1B) are rejected (return false).
+     */
+    fun bankWithdraw(guildId: UUID, actorId: UUID, amount: Long, reason: String): Boolean
 
-    /** Deposit [amount] into [guildId]'s bank as [actorId]. True on success. */
-    fun bankDeposit(guildId: UUID, actorId: UUID, amount: Int, reason: String): Boolean
+    /**
+     * Deposit [amount] into [guildId]'s bank as [actorId]. True on success.
+     * Same Int-bounded constraint as [bankWithdraw].
+     */
+    fun bankDeposit(guildId: UUID, actorId: UUID, amount: Long, reason: String): Boolean
 }
 
 /**
